requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# MLOps
mlflow>=2.9.0
dvc>=3.30.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Monitoring
evidently>=0.4.0

# Technical indicators
ta>=0.11.0

# Visualization & EDA
matplotlib>=3.7.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Linting
flake8>=6.1.0
black>=23.0.0

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
